\name{rectLookup-methods}
\docType{methods}
\alias{rectLookup-methods}
\alias{rectLookup,QuadTree-method}
\title{ Methods for Function \code{rectLookup} in Package \pkg{SearchTrees}}
\description{
 Methods for function \code{rectLookup} in package \pkg{SearchTrees}
}
\section{Methods}{
\describe{

\item{\code{signature(tree = "QuadTree")}}{
}
}}
\keyword{methods}
